+++ docs/changelog.rst
@@ -4,10 +4,17 @@ Changelog
 Releases
 --------
 
+* `Release 0.0.5 - November 6, 2011`_
 * `Release 0.0.4 - August 21, 2011`_
 * `Release 0.0.3 - August 13, 2011`_
 * `Release 0.0.2 - August 13, 2011`_
 * `Release 0.0.1 - August 7, 2011`_
+
+Release 0.0.5 - November 6, 2011
+--------------------------------
+
+* Removed explicit dependency on distribute version 0.6.14.
+* Switch from httplib2 to requests.
 
 Release 0.0.4 - August 21, 2011
 -------------------------------